requirements.txt
django==2.0.5
djangorestframework==3.8.2
Django-environ==0.4.4
django-filter
bootstrap-admin
drf-yasg
swagger_spec_validator
flex